refactor: determineTranslationLoader
